requirements.txt
# LLM API
google-generativeai>=0.3.0

# MCP SDK (official)
mcp>=1.0.0

# Core dependencies
httpx>=0.27.0
pydantic>=2.0.0
anyio>=4.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
